+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - [#2013](https://github.com/teambit/bit/issues/2013) fix bit import when one module resolution alias is a directory of another alias
 - [#2004](https://github.com/teambit/bit/issues/2004) ask for approval before exporting a component to another scope (fork)
-- [#1956](https://github.com/teambit/bit/issues/1956) introduce a new flag `--rewire-scope` for `bit export` to replace the import/require statements in the source to the newly exported scope
+- [#1956](https://github.com/teambit/bit/issues/1956) introduce a new flag `--rewire` for `bit export` to replace the import/require statements in the source to the newly exported scope
 - fix "Converting circular structure to JSON" error when logging a circular metadata object
 - fix exporting to a different scope than workspace configuration of `defaultScope`
 - fix exporting components with and without scope at the same time